refactor: rough-edges hardening for v0.8.27 release prep
Extract RunManager, split connections rail, add install/ARP visibility, and document dev/frozen isolation before tagging.
