+++ .env.example
@@ -10,6 +10,10 @@
 # into BAKLOG-Data:
 # BAKLOG_DATA_DIR=%LOCALAPPDATA%\BAKLOG-Dev
 #
+# Use a different port so browser localStorage is not shared with BAKLOG.exe
+# (both default to http://127.0.0.1:8765 otherwise):
+# PORT=8766
+#
 # Portable zip / thumb drive: add portable.txt beside BAKLOG.exe, or set:
 # BAKLOG_PORTABLE=1
