+++ index.html
@@ -8,7 +8,7 @@
   <!-- Must match pyproject.toml + package.json (see CHANGELOG release discipline). Read at runtime by js/error-boundary.js for bug-bundle metadata. -->
   <!-- Optional override for bug report endpoint during local testing: content="http://127.0.0.1:3000/api/report" -->
   <!-- <meta name="baklog-report-endpoint" content="https://baklog.app/api/report" /> -->
-  <meta name="baklog-version" content="0.8.26" />
+  <meta name="baklog-version" content="0.8.27" />
   <meta name="theme-color" content="#0f172a" />
   <link rel="icon" href="favicon.svg" type="image/svg+xml" />
   <script>
