+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "steam-backlog",
-  "version": "0.8.26",
+  "version": "0.8.27",
   "license": "MIT",
   "private": true,
   "type": "module",
