+++ server.py
@@ -47,8 +47,6 @@ import subprocess
 import sys
 import threading
 import time
-import uuid
-from collections import deque
 from functools import partial
 from http import HTTPStatus
 from http.server import SimpleHTTPRequestHandler, ThreadingHTTPServer
@@ -63,16 +61,13 @@ from shared.built_frontend import (
     maybe_serve_built_index as _maybe_serve_built_index,
 )
 from shared.dev_server_pids import (
-    pid_alive as _pid_alive,
-)
-from shared.dev_server_pids import (
     reclaim_or_exit as _reclaim_or_exit,
 )
 from shared.dev_server_pids import (
     remove_own_pid_file as _remove_own_pid_file,
 )
 from shared.dev_server_pids import (
-    terminate_pid as _terminate_pid,
+    terminate_pid as _terminate_pid_impl,
 )
 from shared.dev_server_pids import (
     write_pid_file as _write_pid_file_impl,
@@ -237,7 +232,7 @@ from shared.server_static import (  # noqa: E402
 from shared.server_static import (  # noqa: E402
     static_class as _static_class_impl,
 )
-from shared.subprocess_guard import _max_run_seconds_from_env, popen_fetcher  # noqa: E402
+from shared.subprocess_guard import _max_run_seconds_from_env  # noqa: E402
 
 MAX_RUN_SECONDS = _max_run_seconds_from_env()
 
@@ -818,8 +813,25 @@ def _read_optional_json(path: Path) -> dict[str, Any] | None:
 INTERNAL_JOBS: dict[str, dict[str, Any]] = _load_internal_jobs()
 
 
+
+from shared.run_manager import (  # noqa: E402
+    Run,
+    RunManager,
+    _load_run_history,
+    _load_run_history_from,  # noqa: F401 — re-exported for tests
+    _run_id_active_on_disk,
+)
+
+MANAGER = RunManager(restore_durable=False)
+
+
+def _terminate_pid(pid: int) -> None:
+    """Default kill hook; tests monkeypatch this on the server module."""
+    _terminate_pid_impl(pid)
+
+
 def _kill_pids_async(pids: list[int]) -> None:
-    """Kill subprocess trees on a daemon thread so HTTP cancel handlers return immediately."""
+    """Default async kill hook; tests monkeypatch this on the server module."""
     unique = list(dict.fromkeys(p for p in pids if p > 0))
     if not unique:
         return
@@ -829,1535 +841,6 @@ def _kill_pids_async(pids: list[int]) -> None:
             _terminate_pid(pid)
 
     threading.Thread(target=_work, name="run-kill", daemon=True).start()
-
-
-def _read_json_file(path: Path, default: Any) -> Any:
-    if not path.exists():
-        return default
-    try:
-        return json.loads(path.read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError):
-        return default
-
-
-def _write_json_atomic(path: Path, payload: Any) -> None:
-    path.parent.mkdir(parents=True, exist_ok=True)
-    tmp = path.with_suffix(path.suffix + ".tmp")
-    tmp.write_text(json.dumps(payload, ensure_ascii=False, indent=2), encoding="utf-8")
-    os.replace(tmp, path)
-
-
-def _read_active_runs() -> list[dict[str, Any]]:
-    with _runs_file_lock:
-        data = _read_json_file(ACTIVE_RUNS_FILE, {"runs": []})
-    runs = data.get("runs") if isinstance(data, dict) else []
-    return runs if isinstance(runs, list) else []
-
-
-def _write_active_runs(runs: list[dict[str, Any]]) -> None:
-    with _runs_file_lock:
-        _write_json_atomic(ACTIVE_RUNS_FILE, {"runs": runs})
-
-
-def _run_id_active_on_disk(run_id: str) -> bool:
-    return any(entry.get("id") == run_id for entry in _read_active_runs())
-
-
-def _fetcher_is_enrich(key: str) -> bool:
-    return FETCHERS.get(key, {}).get("group") == "enrich"
-
-
-def _filter_runs_by_lane(runs: list[Run], lane: str | None) -> list[Run]:
-    """Restrict a run list to one lane. lane="fetcher" drops internal/enrich runs;
-    lane="enrich" keeps enrich only; lane="internal" keeps admin jobs; None = all."""
-    if lane == "fetcher":
-        return [r for r in runs if not r._internal and not r._enrich]
-    if lane == "enrich":
-        return [r for r in runs if r._enrich]
-    if lane == "internal":
-        return [r for r in runs if r._internal]
-    return list(runs)
-
-
-def _load_durable_queue() -> list[dict[str, Any]]:
-    with _runs_file_lock:
-        data = _read_json_file(QUEUE_FILE, {"runs": []})
-    runs = data.get("runs") if isinstance(data, dict) else []
-    return runs if isinstance(runs, list) else []
-
-
-def _save_durable_queue(entries: list[dict[str, Any]]) -> None:
-    with _runs_file_lock:
-        _write_json_atomic(QUEUE_FILE, {"runs": entries})
-
-
-def _load_run_history_from(path: Path | None = None) -> list[dict[str, Any]]:
-    hist_path = path or RUN_HISTORY_FILE
-    data = _read_json_file(hist_path, [])
-    return data if isinstance(data, list) else []
-
-
-def _load_run_history() -> list[dict[str, Any]]:
-    return _load_run_history_from(RUN_HISTORY_FILE)
-
-
-def _save_run_history_to(path: Path, entries: list[dict[str, Any]]) -> None:
-    _write_json_atomic(path, entries[:MAX_HISTORY])
-
-
-def _save_run_history(entries: list[dict[str, Any]]) -> None:
-    _save_run_history_to(RUN_HISTORY_FILE, entries)
-
-
-class Run:
-    """A single queued/running/completed fetcher invocation."""
-
-    __slots__ = (
-        "id", "key", "label", "status", "started_at", "ended_at", "exit_code",
-        "lines", "_lock", "_listeners", "_finished", "_proc", "cancelled", "refresh",
-        "_log_path", "_runs_dir", "profile_id", "_cancelling_since", "_no_proc_since",
-        "_history_note", "_next_seq", "_total_lines", "_finalized",
-        "_internal", "_internal_extra_args", "_enrich",
-    )
-
-    def __init__(
-        self,
-        key: str,
-        refresh: bool = False,
-        *,
-        runs_dir: Path = RUNS_DIR,
-        profile_id: str | None = None,
-        internal: bool = False,
-        enrich: bool = False,
-        extra_args: list[str] | None = None,
-    ) -> None:
-        if internal:
-            if key not in INTERNAL_JOBS:
-                raise KeyError(key)
-            spec = INTERNAL_JOBS[key]
-        else:
-            spec = FETCHERS[key]
-        self.id: str = uuid.uuid4().hex[:12]
-        self.key: str = key
-        self.profile_id: str = profile_id or get_active_profile_id()
-        self.label: str = spec["label"]
-        self.refresh: bool = refresh
-        self._internal = internal
-        self._enrich = enrich and not internal
-        self._internal_extra_args = list(extra_args or [])
-        self.status: str = "queued"  # queued | launching | running | cancelling | done | failed | cancelled
-        self.started_at: float | None = None
-        self.ended_at: float | None = None
-        self.exit_code: int | None = None
-        self.cancelled: bool = False
-        self._finalized: bool = False
-        self._proc: subprocess.Popen[str] | None = None
-        self._runs_dir = runs_dir
-        self._runs_dir.mkdir(parents=True, exist_ok=True)
-        self._log_path = self._runs_dir / f"{self.id}.jsonl"
-        # Ring buffer for live listeners; full log is on disk for replay.
-        self.lines: deque[dict[str, Any]] = deque(maxlen=MAX_LINES_PER_RUN)
-        self._next_seq = 0
-        self._total_lines = 0
-        self._lock = threading.Lock()
-        self._listeners: set[queue.Queue] = set()
-        self._finished = threading.Event()
-        self._cancelling_since: float | None = None
-        self._no_proc_since: float | None = None
-        self._history_note: str | None = None
-        self._restore_seq_from_disk()
-
-    def _restore_seq_from_disk(self) -> None:
-        """Resume monotonic seq / line totals from an existing jsonl log."""
-        if not self._log_path.exists():
-            return
-        max_seq = 0
-        count = 0
-        fallback = 0
-        try:
-            with self._log_path.open(encoding="utf-8") as f:
-                for line in f:
-                    line = line.strip()
-                    if not line:
-                        continue
-                    count += 1
-                    msg = json.loads(line)
-                    seq = msg.get("seq")
-                    if seq is None:
-                        fallback += 1
-                        seq = fallback
-                    max_seq = max(max_seq, int(seq))
-        except (OSError, json.JSONDecodeError) as exc:
-            print(f"[runs] log scan failed {self.id}: {exc!r}", file=sys.stderr)
-            return
-        self._next_seq = max_seq
-        self._total_lines = count
-
-    def to_summary(self) -> dict[str, Any]:
-        summary: dict[str, Any] = {
-            "id": self.id,
-            "key": self.key,
-            "label": self.label,
-            "status": self.status,
-            "started_at": self.started_at,
-            "ended_at": self.ended_at,
-            "exit_code": self.exit_code,
-            "line_count": self._total_lines,
-        }
-        if self.exit_code == 4:
-            summary["failure_kind"] = "auth"
-        if self._history_note:
-            summary["note"] = self._history_note
-        summary["profile_id"] = self.profile_id
-        if self._enrich:
-            summary["lane"] = "enrich"
-        elif self._internal:
-            summary["lane"] = "internal"
-        else:
-            summary["lane"] = "fetcher"
-        if not self._internal:
-            summary["group"] = FETCHERS.get(self.key, {}).get("group")
-        return summary
-
-    def add_line(self, stream: str, text: str) -> None:
-        with self._lock:
-            self._next_seq += 1
-            self._total_lines += 1
-            seq = self._next_seq
-            msg = {
-                "seq": seq,
-                "t": time.time(),
-                "stream": stream,
-                "text": _redact_log_line(text),
-            }
-            self.lines.append(msg)
-            try:
-                with self._log_path.open("a", encoding="utf-8") as f:
-                    f.write(json.dumps(msg, ensure_ascii=False) + "\n")
-            except OSError as exc:
-                print(f"[runs] log write failed {self.id}: {exc!r}", file=sys.stderr)
-            for q in list(self._listeners):
-                try:
-                    q.put_nowait(("line", msg))
-                except queue.Full:
-                    # Drop slow listeners rather than block the worker thread.
-                    self._listeners.discard(q)
-
-    def broadcast(self, event: str, data: dict[str, Any]) -> None:
-        with self._lock:
-            for q in list(self._listeners):
-                try:
-                    q.put_nowait((event, data))
-                except queue.Full:
-                    self._listeners.discard(q)
-
-    def replay_lines(self, since: int = 0) -> list[dict[str, Any]]:
-        """Return log lines with seq > since (full log on disk; ring buffer fallback)."""
-        since = max(0, int(since))
-
-        def _with_seq(messages: list[dict[str, Any]]) -> list[dict[str, Any]]:
-            out: list[dict[str, Any]] = []
-            fallback = 0
-            for msg in messages:
-                seq = msg.get("seq")
-                if seq is None:
-                    fallback += 1
-                    msg = {**msg, "seq": fallback}
-                out.append(msg)
-            return [m for m in out if int(m.get("seq", 0)) > since]
-
-        if self._log_path.exists():
-            replay: list[dict[str, Any]] = []
-            try:
-                with self._log_path.open(encoding="utf-8") as f:
-                    for line in f:
-                        line = line.strip()
-                        if not line:
-                            continue
-                        replay.append(json.loads(line))
-                return _with_seq(replay)
-            except (OSError, json.JSONDecodeError) as exc:
-                print(f"[runs] log read failed {self.id}: {exc!r}", file=sys.stderr)
-        with self._lock:
-            return _with_seq(list(self.lines))
-
-    def attach_listener(self, since: int = 0) -> tuple[queue.Queue, list[dict[str, Any]], bool]:
-        """Return (queue, replay-buffer, already-finished)."""
-        q: queue.Queue = queue.Queue(maxsize=1024)
-        with self._lock:
-            replay = self.replay_lines(since)
-            done = self._finished.is_set()
-            if not done:
-                self._listeners.add(q)
-        return q, replay, done
-
-    def detach_listener(self, q: queue.Queue) -> None:
-        with self._lock:
-            self._listeners.discard(q)
-
-    def mark_finished(self) -> None:
-        self._finished.set()
-
-    def argv(self) -> list[str]:
-        if self._internal:
-            return _internal_job_argv(
-                INTERNAL_JOBS[self.key],
-                self._internal_extra_args,
-            )
-        spec = FETCHERS[self.key]
-        argv = list(spec["argv"])
-        if self.refresh:
-            for arg in spec.get("refreshArgs") or []:
-                if arg not in argv:
-                    argv.append(arg)
-        return argv
-
-    def cancel(self) -> tuple[bool, list[int]]:
-        """Mark cancelled/cancelling. Returns (changed, pids_to_kill). Never kills inline.
-
-        broadcast()/mark_finished() acquire self._lock, so all notifications are
-        emitted AFTER releasing the lock here — Run._lock is a plain (non-reentrant)
-        Lock and broadcasting under it deadlocks the calling (HTTP) thread.
-        """
-        proc = None
-        notify_done = False
-        notify_cancelling = False
-        with self._lock:
-            if self.status in ("done", "failed", "cancelled") or self._finished.is_set():
-                return False, []
-            if self.status == "queued":
-                self.status = "cancelled"
-                self.exit_code = -1
-                self.ended_at = time.time()
-                notify_done = True
-            elif self.status in ("launching", "running"):
-                self.cancelled = True
-                proc = self._proc
-                if proc is None or proc.poll() is not None:
-                    self.status = "cancelled"
-                    self.exit_code = -1
-                    self.ended_at = time.time()
-                    notify_done = True
-                else:
-                    self.status = "cancelling"
-                    self._cancelling_since = time.monotonic()
-                    notify_cancelling = True
-            elif self.status == "cancelling":
-                return False, []
-            else:
-                return False, []
-        if notify_cancelling:
-            self.broadcast("status", {"status": self.status, "started_at": self.started_at})
-        if notify_done:
-            self.add_line("stderr", "[server] cancelled before start")
-            self.mark_finished()
-            self.broadcast(
-                "done",
-                {
-                    "status": self.status,
-                    "exit_code": self.exit_code,
-                    "started_at": self.started_at,
-                    "ended_at": self.ended_at,
-                },
-            )
-            return True, []
-        pids: list[int] = []
-        if proc is not None and proc.poll() is None and proc.pid:
-            pids.append(proc.pid)
-        return True, pids
-
-
-class RunManager:
-    """Single-worker queue. Fetchers may share locks (PSN session, etc.) so
-    we deliberately serialize them rather than spawn in parallel."""
-
-    def __init__(
-        self,
-        runs_dir: Path | None = None,
-        *,
-        enable_watchdog: bool = True,
-        reap_orphans: bool | None = None,
-        restore_durable: bool = True,
-    ) -> None:
-        self._runs_dir = runs_dir or RUNS_DIR
-        self._runs_dir.mkdir(parents=True, exist_ok=True)
-        self._lock = threading.Lock()
-        # Three parallel lanes, each cap=1: fetcher (library/wishlist/prices),
-        # enrich (HLTB/reviews/covers/tags), and internal (admin jobs).
-        self._queue: queue.Queue[Run] = queue.Queue()
-        self._enrich_queue: queue.Queue[Run] = queue.Queue()
-        self._internal_queue: queue.Queue[Run] = queue.Queue()
-        self._pending: list[Run] = []  # queued + active (all lanes), submission order
-        self._history: deque[dict[str, Any]] = deque(
-            _load_run_history_from(self._runs_dir / "history.json")[-MAX_HISTORY:],
-            maxlen=MAX_HISTORY,
-        )
-        self._active: Run | None = None
-        self._enrich_active: Run | None = None
-        self._internal_active: Run | None = None
-        self._runs_by_id: dict[str, Run] = {}
-        self._last_queue_kick_at = 0.0
-        self._watchdog_stop = threading.Event()
-        # Only reap on explicit RunManager(runs_dir=...) for tests, or when the
-        # dev server calls MANAGER._reap_orphan_processes() at boot. Importing
-        # server.py (e.g. pytest) must not kill live fetchers from a running dev
-        # server — that shared active.json lives on disk.
-        if reap_orphans if reap_orphans is not None else runs_dir is not None:
-            self._reap_orphan_processes()
-        self._start_worker_thread()
-        self._watchdog_thread: threading.Thread | None = None
-        if enable_watchdog:
-            self._watchdog_thread = threading.Thread(
-                target=self._watchdog_loop, name="run-watchdog", daemon=True
-            )
-            self._watchdog_thread.start()
-        # Production server defers restore to main() after bind; tests pass True.
-        if restore_durable:
-            self._restore_durable_queue()
-
-    def _start_worker_thread(self) -> None:
-        self._worker_thread = threading.Thread(
-            target=self._worker_loop, args=("fetcher",), name="run-worker", daemon=True
-        )
-        self._worker_thread.start()
-        self._enrich_worker_thread = threading.Thread(
-            target=self._worker_loop, args=("enrich",), name="run-worker-enrich", daemon=True
-        )
-        self._enrich_worker_thread.start()
-        self._internal_worker_thread = threading.Thread(
-            target=self._worker_loop, args=("internal",), name="run-worker-internal", daemon=True
-        )
-        self._internal_worker_thread.start()
-
-    def _ensure_worker_thread(self) -> None:
-        """Restart any lane worker if its daemon thread died."""
-        if not self._worker_thread.is_alive():
-            print("[runs] fetcher worker thread died — restarting", file=sys.stderr, flush=True)
-            self._worker_thread = threading.Thread(
-                target=self._worker_loop, args=("fetcher",), name="run-worker", daemon=True
-            )
-            self._worker_thread.start()
-        enrich = getattr(self, "_enrich_worker_thread", None)
-        if enrich is None or not enrich.is_alive():
-            print("[runs] enrich worker thread died — restarting", file=sys.stderr, flush=True)
-            self._enrich_worker_thread = threading.Thread(
-                target=self._worker_loop, args=("enrich",), name="run-worker-enrich", daemon=True
-            )
-            self._enrich_worker_thread.start()
-        internal = getattr(self, "_internal_worker_thread", None)
-        if internal is None or not internal.is_alive():
-            print("[runs] internal worker thread died — restarting", file=sys.stderr, flush=True)
-            self._internal_worker_thread = threading.Thread(
-                target=self._worker_loop, args=("internal",), name="run-worker-internal", daemon=True
-            )
-            self._internal_worker_thread.start()
-
-    def _resync_stalled_queue(self) -> int:
-        """Put pending queued runs back on the worker queue when nothing is active.
-
-        This heals the wedge where runs sit in ``_pending`` with status ``queued``
-        but were never handed to ``_queue.get()`` (typically after the worker thread
-        exited while the queue was empty). Runs both lanes independently.
-        """
-        return (
-            self._resync_lane("fetcher")
-            + self._resync_lane("enrich")
-            + self._resync_lane("internal")
-        )
-
-    def _lane_queue(self, lane: str) -> queue.Queue[Run]:
-        if lane == "internal":
-            return self._internal_queue
-        if lane == "enrich":
-            return self._enrich_queue
-        return self._queue
-
-    def _lane_active(self, lane: str) -> Run | None:
-        if lane == "internal":
-            return self._internal_active
-        if lane == "enrich":
-            return self._enrich_active
-        return self._active
-
-    def _set_lane_active(self, lane: str, run: Run | None) -> None:
-        if lane == "internal":
-            self._internal_active = run
-        elif lane == "enrich":
-            self._enrich_active = run
-        else:
-            self._active = run
-
-    def _run_in_lane(self, run: Run, lane: str) -> bool:
-        if lane == "internal":
-            return run._internal
-        if lane == "enrich":
-            return run._enrich
-        return not run._internal and not run._enrich
-
-    def _resync_lane(self, lane: str) -> int:
-        lane_queue = self._lane_queue(lane)
-        to_put: list[Run] = []
-        with self._lock:
-            active = self._lane_active(lane)
-            if active is not None and active._finished.is_set():
-                self._set_lane_active(lane, None)
-                active = None
-            if active is not None:
-                return 0
-            if lane_queue.qsize() > 0:
-                return 0
-            for r in self._pending:
-                if not self._run_in_lane(r, lane):
-                    continue
-                if r.status == "queued" and not r._finished.is_set():
-                    to_put.append(r)
-        for r in to_put:
-            lane_queue.put(r)
-        if to_put:
-            keys = ", ".join(r.key for r in to_put)
-            print(
-                f"[runs] re-queued {len(to_put)} stalled {lane} run(s): {keys}",
-                file=sys.stderr,
-                flush=True,
-            )
-        return len(to_put)
-
-    def _kick_queue_if_stalled(self) -> None:
-        self._ensure_worker_thread()
-        self._resync_stalled_queue()
-
-    def _kick_queue_if_stalled_throttled(self) -> None:
-        now = time.monotonic()
-        if now - self._last_queue_kick_at < 1.0:
-            return
-        self._last_queue_kick_at = now
-        self._kick_queue_if_stalled()
-
-    def _watchdog_loop(self) -> None:
-        while not self._watchdog_stop.wait(WATCHDOG_INTERVAL_SEC):
-            try:
-                self._kick_queue_if_stalled()
-                self._force_finalize_stuck_cancelling()
-                self._force_finalize_orphaned_runs()
-            except Exception as exc:  # noqa: BLE001
-                print(f"[runs] watchdog error: {exc!r}", file=sys.stderr, flush=True)
-
-    def _force_finalize_stuck_cancelling(self) -> None:
-        now = time.monotonic()
-        stuck: list[Run] = []
-        with self._lock:
-            for r in self._pending:
-                if (
-                    r.status == "cancelling"
-                    and r._cancelling_since is not None
-                    and now - r._cancelling_since > CANCEL_STUCK_GRACE_SEC
-                ):
-                    stuck.append(r)
-            for active in (self._active, self._enrich_active, self._internal_active):
-                if (
-                    active
-                    and active.status == "cancelling"
-                    and active._cancelling_since is not None
-                    and now - active._cancelling_since > CANCEL_STUCK_GRACE_SEC
-                    and active not in stuck
-                ):
-                    stuck.append(active)
-        for run in stuck:
-            pids = self._collect_pids_for_run(run, [])
-            if pids:
-                _kill_pids_async(pids)
-            with run._lock:
-                if not run._finished.is_set():
-                    run.status = "cancelled"
-                    run.exit_code = -1
-                    if run.ended_at is None:
-                        run.ended_at = time.time()
-            if not run._finished.is_set():
-                run.mark_finished()
-                run.broadcast(
-                    "done",
-                    {
-                        "status": run.status,
-                        "exit_code": run.exit_code,
-                        "started_at": run.started_at,
-                        "ended_at": run.ended_at,
-                    },
-                )
-            self._finalize_run(run)
-
-    def _run_has_live_process(self, run: Run) -> bool:
-        if run._proc is not None and run._proc.poll() is None:
-            return True
-        for entry in _read_active_runs():
-            if entry.get("id") == run.id:
-                pid = int(entry.get("pid") or 0)
-                if _pid_alive(pid):
-                    return True
-        return False
-
-    def _force_finalize_orphaned_runs(self) -> None:
-        """Force-finalize launching/running runs with no live subprocess."""
-        now = time.monotonic()
-        stuck: list[Run] = []
-        with self._lock:
-            candidates: list[Run] = []
-            for active in (self._active, self._enrich_active, self._internal_active):
-                if (
-                    active
-                    and active.status in ("launching", "running")
-                    and not active._finished.is_set()
-                ):
-                    candidates.append(active)
-            for r in self._pending:
-                if (
-                    r.status in ("launching", "running")
-                    and not r._finished.is_set()
-                    and r not in candidates
-                ):
-                    candidates.append(r)
-            for run in candidates:
-                if self._run_has_live_process(run):
-                    run._no_proc_since = None
-                    continue
-                if run._no_proc_since is None:
-                    run._no_proc_since = now
-                    continue
-                if now - run._no_proc_since > STUCK_NO_PROC_GRACE_SEC:
-                    stuck.append(run)
-        for run in stuck:
-            pids = self._collect_pids_for_run(run, [])
-            if pids:
-                _kill_pids_async(pids)
-            with run._lock:
-                if not run._finished.is_set():
-                    run.status = "failed"
-                    run.exit_code = -1
-                    if run.ended_at is None:
-                        run.ended_at = time.time()
-                    run._history_note = (
-                        "force-finalized: no live subprocess (worker stalled)"
-                    )
-            if not run._finished.is_set():
-                run.add_line(
-                    "stderr",
-                    "[server] no live subprocess — force-finalizing stalled run "
-                    "so the queue can advance",
-                )
-                run.mark_finished()
-                run.broadcast(
-                    "done",
-                    {
-                        "status": run.status,
-                        "exit_code": run.exit_code,
-                        "started_at": run.started_at,
-                        "ended_at": run.ended_at,
-                    },
-                )
-            self._finalize_run(run)
-
-    def _collect_pids_for_run(self, run: Run, proc_pids: list[int]) -> list[int]:
-        pids = list(proc_pids)
-        if run._proc is not None and run._proc.poll() is None and run._proc.pid:
-            if run._proc.pid not in pids:
-                pids.append(run._proc.pid)
-        for entry in _read_active_runs():
-            if entry.get("id") == run.id:
-                pid = int(entry.get("pid") or 0)
-                if pid > 0 and pid not in pids:
-                    pids.append(pid)
-        return pids
-
-    def _complete_cancel_after_kill(self, run: Run) -> None:
-        proc = run._proc
-        if proc is not None and proc.poll() is None:
-            try:
-                proc.wait(timeout=TERMINATE_GRACE_SEC)
-            except Exception:  # noqa: BLE001
-                pass
-        # mark_finished()/broadcast() acquire run._lock themselves, so we must
-        # NOT hold it here — Run._lock is a plain (non-reentrant) Lock.
-        should_finish = False
-        with run._lock:
-            if not run._finished.is_set():
-                run.status = "cancelled"
-                run.exit_code = -1
-                if run.ended_at is None:
-                    run.ended_at = time.time()
-                should_finish = True
-        if should_finish:
-            run.mark_finished()
-            run.broadcast(
-                "done",
-                {
-                    "status": run.status,
-                    "exit_code": run.exit_code,
-                    "started_at": run.started_at,
-                    "ended_at": run.ended_at,
-                },
-            )
-        self._finalize_run(run)
-
-    def stream_terminal_summary(self, run_id: str) -> dict[str, Any] | None:
-        """History summary for a run no longer in memory (SSE reconnect after finish)."""
-        with self._lock:
-            for h in self._history:
-                if h.get("id") == run_id:
-                    return h
-        hist_file = self._runs_dir / "history.json"
-        for h in _load_run_history_from(hist_file):
-            if h.get("id") == run_id:
-                return h
-        return None
-
-    def rebind_profile_paths(self) -> None:
-        """Point run storage at the active profile after POST /api/profiles/active."""
-        self._runs_dir = runs_dir()
-        self._runs_dir.mkdir(parents=True, exist_ok=True)
-        active_pid = get_active_profile_id()
-        with self._lock:
-            self._history = deque(
-                _load_run_history_from(self._runs_dir / "history.json")[-MAX_HISTORY:],
-                maxlen=MAX_HISTORY,
-            )
-            self._runs_by_id = {
-                rid: run
-                for rid, run in self._runs_by_id.items()
-                if run.profile_id == active_pid
-            }
-            self._pending = [r for r in self._pending if r.profile_id == active_pid]
-            if self._active is not None and self._active.profile_id != active_pid:
-                self._active = None
-            if (
-                self._enrich_active is not None
-                and self._enrich_active.profile_id != active_pid
-            ):
-                self._enrich_active = None
-            if (
-                self._internal_active is not None
-                and self._internal_active.profile_id != active_pid
-            ):
-                self._internal_active = None
-
-    def _persist_queue(self) -> None:
-        with self._lock:
-            entries = [
-                {"id": r.id, "key": r.key, "refresh": r.refresh}
-                for r in self._pending
-                if r.status == "queued"
-            ]
-        _save_durable_queue(entries)
-
-    def _latest_history_for_key(self, key: str) -> dict[str, Any] | None:
-        with self._lock:
-            for h in self._history:
-                if h.get("key") == key:
-                    return h
-        return None
-
-    def _restore_durable_queue(self) -> None:
-        history_ids = {h.get("id") for h in self._history}
-        restored = 0
-        for entry in _load_durable_queue():
-            key = entry.get("key")
-            if not key or key not in FETCHERS:
-                continue
-            if entry.get("id") in history_ids:
-                continue
-            latest = self._latest_history_for_key(key)
-            if latest and latest.get("status") == "done":
-                continue
-            try:
-                self.submit(key, refresh=bool(entry.get("refresh")))
-                restored += 1
-            except ValueError:
-                break
-        if restored:
-            print(f"[runs] restored {restored} queued run(s) from durable queue", file=sys.stderr)
-
-    def _prune_runs_by_id(self) -> None:
-        with self._lock:
-            if len(self._runs_by_id) <= MAX_HISTORY:
-                return
-            keep_ids = {r.id for r in self._pending}
-            keep_ids.update(h.get("id") for h in self._history if h.get("id"))
-            for rid in list(self._runs_by_id):
-                if rid not in keep_ids:
-                    del self._runs_by_id[rid]
-
-    def _append_history(self, summary: dict[str, Any], *, profile_id: str) -> None:
-        hist_file = runs_dir(profile_id=profile_id) / "history.json"
-        entries = _load_run_history_from(hist_file)
-        entries.insert(0, summary)
-        _save_run_history_to(hist_file, entries)
-        with self._lock:
-            if profile_id == get_active_profile_id():
-                self._history.appendleft(summary)
-                while len(self._history) > MAX_HISTORY:
-                    self._history.pop()
-
-    def _register_pending_run(self, run: Run) -> None:
-        """Cross-process hint so stream handlers can wait for another dev server."""
-        entry = {
-            "id": run.id,
-            "pid": 0,
-            "key": run.key,
-            "label": run.label,
-            "started_at": run.started_at or time.time(),
-        }
-        active = [e for e in _read_active_runs() if e.get("id") != run.id]
-        active.append(entry)
-        _write_active_runs(active)
-
-    def _register_active_process(self, run: Run, pid: int) -> None:
-        entry = {
-            "id": run.id,
-            "pid": pid,
-            "key": run.key,
-            "label": run.label,
-            "started_at": run.started_at,
-        }
-        active = [e for e in _read_active_runs() if e.get("id") != run.id]
-        active.append(entry)
-        _write_active_runs(active)
-
-    def _unregister_active_process(self, run_id: str) -> None:
-        active = [e for e in _read_active_runs() if e.get("id") != run_id]
-        _write_active_runs(active)
-
-    def _reap_orphan_processes(self) -> None:
-        for entry in _read_active_runs():
-            pid = int(entry.get("pid") or 0)
-            if _pid_alive(pid):
-                _terminate_pid(pid)
-            summary = {
-                "id": entry.get("id", "?"),
-                "key": entry.get("key", "?"),
-                "label": entry.get("label", "?"),
-                "status": "failed",
-                "started_at": entry.get("started_at"),
-                "ended_at": time.time(),
-                "exit_code": -1,
-                "line_count": 0,
-                "note": "orphaned — previous server stopped while this fetcher was running",
-            }
-            self._append_history(summary, profile_id=get_active_profile_id())
-        _write_active_runs([])
-
-    def shutdown(self) -> None:
-        self._watchdog_stop.set()
-        with self._lock:
-            pending = list(self._pending)
-            actives = [self._active, self._enrich_active, self._internal_active]
-        kill_pids: list[int] = []
-        for run in pending:
-            changed, pids = run.cancel()
-            if changed:
-                kill_pids.extend(self._collect_pids_for_run(run, pids))
-        for active in actives:
-            if active is not None:
-                changed, pids = active.cancel()
-                if changed:
-                    kill_pids.extend(self._collect_pids_for_run(active, pids))
-        if kill_pids:
-            for pid in dict.fromkeys(kill_pids):
-                _terminate_pid(pid)
-        _write_active_runs([])
-        _save_durable_queue([])
-        # Wake both worker lanes so their blocking queue.get() returns and the
-        # threads exit, letting join_threads() finish immediately rather than
-        # waiting out each 5s join timeout.
-        for lane_queue in (self._queue, self._enrich_queue, self._internal_queue):
-            try:
-                lane_queue.put_nowait(None)
-            except Exception:  # noqa: BLE001 - best-effort wakeup
-                pass
-        self.join_threads(timeout=5.0)
-
-    def join_threads(self, timeout: float = 5.0) -> None:
-        """Stop watchdog and wait for worker/watchdog threads (bounded)."""
-        self._watchdog_stop.set()
-        wt = getattr(self, "_watchdog_thread", None)
-        if wt is not None and wt.is_alive():
-            wt.join(timeout=timeout)
-        if self._worker_thread.is_alive():
-            self._worker_thread.join(timeout=timeout)
-        enrich = getattr(self, "_enrich_worker_thread", None)
-        if enrich is not None and enrich.is_alive():
-            enrich.join(timeout=timeout)
-        internal = getattr(self, "_internal_worker_thread", None)
-        if internal is not None and internal.is_alive():
-            internal.join(timeout=timeout)
-
-    def submit(self, key: str, *, refresh: bool = False) -> Run:
-        if key not in FETCHERS:
-            raise KeyError(key)
-        is_enrich = _fetcher_is_enrich(key)
-
-        def _in_lane(r: Run) -> bool:
-            if is_enrich:
-                return r._enrich
-            return not r._internal and not r._enrich
-
-        with self._lock:
-            active = self._enrich_active if is_enrich else self._active
-            if active and active.key == key and active.status in _IN_FLIGHT_STATUSES:
-                raise ValueError(f"{key} already queued or running")
-            if any(
-                _in_lane(r) and r.key == key and r.status in _IN_FLIGHT_STATUSES
-                for r in self._pending
-            ):
-                raise ValueError(f"{key} already queued or running")
-            in_flight = sum(
-                1 for r in self._pending if _in_lane(r) and r.status in _IN_FLIGHT_STATUSES
-            )
-            if (
-                active
-                and active.status in _IN_FLIGHT_STATUSES
-                and active not in self._pending
-            ):
-                in_flight += 1
-            if in_flight >= 1:
-                lane_label = "enrich" if is_enrich else "fetch"
-                raise ValueError(
-                    f"queue full — a {lane_label} is already running; "
-                    "wait for it to finish before starting another"
-                )
-            profile_id = get_active_profile_id()
-            run = Run(
-                key,
-                refresh=refresh,
-                runs_dir=runs_dir(profile_id=profile_id),
-                profile_id=profile_id,
-                enrich=is_enrich,
-            )
-            self._pending.append(run)
-            self._runs_by_id[run.id] = run
-            (self._enrich_queue if is_enrich else self._queue).put(run)
-        self._register_pending_run(run)
-        self._persist_queue()
-        self._ensure_worker_thread()
-        return run
-
-    def submit_internal(
-        self,
-        key: str,
-        extra_args: list[str] | None = None,
-        *,
-        profile_id: str | None = None,
-    ) -> Run:
-        if key not in INTERNAL_JOBS:
-            raise KeyError(key)
-        with self._lock:
-            active = self._internal_active
-            if active and active.key == key and active.status in _IN_FLIGHT_STATUSES:
-                raise ValueError(f"{key} already queued or running")
-            if any(
-                r._internal and r.key == key and r.status in _IN_FLIGHT_STATUSES
-                for r in self._pending
-            ):
-                raise ValueError(f"{key} already queued or running")
-            # The internal lane is independent of the fetcher lane: count only
-            # internal in-flight runs so a running library fetch never blocks
-            # admin Publish/Enrich (and vice versa). Internal jobs still
-            # serialize with each other (cap 1) so claimSources/buildClaims
-            # don't race the shared auto feed.
-            in_flight = sum(
-                1 for r in self._pending if r._internal and r.status in _IN_FLIGHT_STATUSES
-            )
-            if (
-                active
-                and active.status in _IN_FLIGHT_STATUSES
-                and active not in self._pending
-            ):
-                in_flight += 1
-            if in_flight >= 1:
-                raise ValueError(
-                    "an admin job is already running; "
-                    "wait for it to finish before starting another"
-                )
-            pid = profile_id or get_active_profile_id()
-            run = Run(
-                key,
-                runs_dir=runs_dir(profile_id=pid),
-                profile_id=pid,
-                internal=True,
-                extra_args=extra_args or [],
-            )
-            self._pending.append(run)
-            self._runs_by_id[run.id] = run
-            self._internal_queue.put(run)
-        self._register_pending_run(run)
-        self._persist_queue()
-        self._ensure_worker_thread()
-        return run
-
-    def cancel(self, run_id: str) -> tuple[Run | None, str | None]:
-        with self._lock:
-            run = self._runs_by_id.get(run_id)
-            if run is None:
-                return None, "not_found"
-            if run.status in ("done", "failed", "cancelled") or run._finished.is_set():
-                return None, "already_finished"
-        changed, proc_pids = run.cancel()
-        if not changed:
-            return None, "already_finished"
-        pids = self._collect_pids_for_run(run, proc_pids)
-        with self._lock:
-            if run in self._pending:
-                self._pending.remove(run)
-        self._persist_queue()
-        if run._finished.is_set():
-            self._finalize_run(run)
-        elif run.status == "cancelling":
-            if pids:
-                _kill_pids_async(pids)
-            self._schedule_cancel_completion(run)
-        return run, None
-
-    def _schedule_cancel_completion(self, run: Run) -> None:
-        if self._worker_thread.is_alive():
-            threading.Thread(
-                target=self._complete_cancel_after_kill,
-                args=(run,),
-                name=f"run-cancel-{run.id}",
-                daemon=True,
-            ).start()
-        else:
-            self._complete_cancel_after_kill(run)
-
-    def cancel_all(
-        self,
-        *,
-        profile_id: str | None = None,
-        lane: str | None = None,
-    ) -> list[dict[str, Any]]:
-        """Cancel queued/running runs. lane="fetcher"/"internal" scopes to one lane.
-
-        The dashboard passes lane="fetcher" so its Cancel button never kills an
-        admin job (buildClaims/claimSources) running in the internal lane.
-        """
-        with self._lock:
-            targets = [
-                r for r in self._pending if r.status in _IN_FLIGHT_STATUSES
-            ]
-            for active in (self._active, self._enrich_active, self._internal_active):
-                if (
-                    active
-                    and active.status in _IN_FLIGHT_STATUSES
-                    and active not in targets
-                ):
-                    targets.append(active)
-        targets = _filter_runs_by_lane(targets, lane)
-        if profile_id is not None:
-            targets = [r for r in targets if r.profile_id == profile_id]
-        all_pids: list[int] = []
-        summaries: list[dict[str, Any]] = []
-        to_finalize_now: list[Run] = []
-        to_complete_async: list[Run] = []
-        for run in targets:
-            changed, proc_pids = run.cancel()
-            if not changed:
-                continue
-            all_pids.extend(self._collect_pids_for_run(run, proc_pids))
-            with self._lock:
-                if run in self._pending:
-                    self._pending.remove(run)
-            if run._finished.is_set():
-                to_finalize_now.append(run)
-            elif run.status == "cancelling":
-                to_complete_async.append(run)
-            summaries.append(run.to_summary())
-        self._persist_queue()
-        if all_pids:
-            _kill_pids_async(all_pids)
-        for run in to_finalize_now:
-            self._finalize_run(run)
-        for run in to_complete_async:
-            self._schedule_cancel_completion(run)
-        return summaries
-
-    def force_reset(
-        self,
-        *,
-        profile_id: str | None = None,
-        lane: str | None = None,
-    ) -> dict[str, Any]:
-        """Kill tracked PIDs, clear queue state, finalize in-flight runs.
-
-        lane="fetcher"/"internal" scopes the reset to a single lane so the
-        dashboard force-reset leaves admin (internal) jobs untouched.
-        """
-        with self._lock:
-            targets = [
-                r for r in self._pending if r.status in _IN_FLIGHT_STATUSES
-            ]
-            for active in (self._active, self._enrich_active, self._internal_active):
-                if (
-                    active
-                    and active.status in _IN_FLIGHT_STATUSES
-                    and active not in targets
-                ):
-                    targets.append(active)
-        targets = _filter_runs_by_lane(targets, lane)
-        if profile_id is not None:
-            targets = [r for r in targets if r.profile_id == profile_id]
-        all_pids: list[int] = []
-        if lane is None:
-            for entry in _read_active_runs():
-                pid = int(entry.get("pid") or 0)
-                if pid <= 0:
-                    continue
-                if profile_id is not None and str(entry.get("profile_id") or "") != str(profile_id):
-                    continue
-                all_pids.append(pid)
-        for run in targets:
-            run.cancelled = True
-            with run._lock:
-                if run.status in _IN_FLIGHT_STATUSES and not run._finished.is_set():
-                    run.status = "cancelled"
-                    run.exit_code = -1
-                    run.ended_at = time.time()
-            all_pids.extend(self._collect_pids_for_run(run, []))
-        target_ids = {run.id for run in targets}
-        drains: list[queue.Queue] = []
-        if lane in (None, "fetcher"):
-            drains.append(self._queue)
-        if lane in (None, "enrich"):
-            drains.append(self._enrich_queue)
-        if lane in (None, "internal"):
-            drains.append(self._internal_queue)
-        for drain in drains:
-            while True:
-                try:
-                    drain.get_nowait()
-                except queue.Empty:
-                    break
-        with self._lock:
-            if lane is None:
-                self._pending.clear()
-                self._active = None
-                self._enrich_active = None
-                self._internal_active = None
-            else:
-                self._pending = [r for r in self._pending if r not in targets]
-                if lane == "fetcher":
-                    self._active = None
-                elif lane == "enrich":
-                    self._enrich_active = None
-                elif lane == "internal":
-                    self._internal_active = None
-        if lane is None:
-            _write_active_runs([])
-        else:
-            _write_active_runs(
-                [e for e in _read_active_runs() if e.get("id") not in target_ids]
-            )
-        _save_durable_queue([])
-        if all_pids:
-            _kill_pids_async(list(dict.fromkeys(all_pids)))
-        summaries: list[dict[str, Any]] = []
-        for run in targets:
-            if not run._finished.is_set():
-                run.mark_finished()
-                run.broadcast(
-                    "done",
-                    {
-                        "status": "cancelled",
-                        "exit_code": run.exit_code or -1,
-                        "started_at": run.started_at,
-                        "ended_at": run.ended_at,
-                    },
-                )
-            self._finalize_run(run)
-            summaries.append(run.to_summary())
-        self._ensure_worker_thread()
-        return {"cancelled": summaries, "force": True}
-
-    def _in_flight_targets(self, profile_id: str | None = None) -> list[Run]:
-        """Queued or running jobs in either lane, including actives dropped from _pending."""
-        with self._lock:
-            targets = [
-                r for r in self._pending if r.status in _IN_FLIGHT_STATUSES
-            ]
-            for active in (self._active, self._enrich_active, self._internal_active):
-                if (
-                    active
-                    and active.status in _IN_FLIGHT_STATUSES
-                    and active not in targets
-                ):
-                    targets.append(active)
-        if profile_id is not None:
-            targets = [r for r in targets if r.profile_id == profile_id]
-        return targets
-
-    def has_runs_for_profile(self, profile_id: str) -> bool:
-        """True if any queued or running job in either lane is bound to this profile."""
-        return bool(self._in_flight_targets(profile_id))
-
-    def cancel_all_and_wait(
-        self,
-        timeout: float = SWITCH_CANCEL_WAIT_SEC,
-    ) -> dict[str, Any]:
-        """Cancel every in-flight job and wait for each to finish (bounded)."""
-        targets = self._in_flight_targets()
-        cancelled: list[dict[str, Any]] = []
-        for run in targets:
-            run_obj, err = self.cancel(run.id)
-            if run_obj is not None and err is None:
-                cancelled.append(run_obj.to_summary())
-        deadline = time.monotonic() + timeout
-        stragglers: list[dict[str, Any]] = []
-        for run in targets:
-            remaining = deadline - time.monotonic()
-            if remaining > 0:
-                run._finished.wait(remaining)
-            if not run._finished.is_set():
-                stragglers.append(run.to_summary())
-        if stragglers:
-            ids = ", ".join(s.get("id", "?") for s in stragglers)
-            print(
-                f"WARN: {len(stragglers)} run(s) still not finished after "
-                f"{timeout}s cancel wait: {ids}",
-                file=sys.stderr,
-                flush=True,
-            )
-        return {"cancelled": cancelled, "stragglers": stragglers}
-
-    def get(self, run_id: str) -> Run | None:
-        with self._lock:
-            return self._runs_by_id.get(run_id)
-
-    def snapshot(self) -> dict[str, Any]:
-        self._kick_queue_if_stalled_throttled()
-        with self._lock:
-            # active/queue cover the fetcher lane; enrich and internal are separate.
-            active = (
-                self._active.to_summary()
-                if self._active and self._active.status in _IN_FLIGHT_STATUSES
-                else None
-            )
-            queued = [
-                r.to_summary()
-                for r in self._pending
-                if not r._internal
-                and not r._enrich
-                and r.status == "queued"
-                and r is not self._active
-            ]
-            enrich_active = (
-                self._enrich_active.to_summary()
-                if self._enrich_active
-                and self._enrich_active.status in _IN_FLIGHT_STATUSES
-                else None
-            )
-            enrich_queue = [
-                r.to_summary()
-                for r in self._pending
-                if r._enrich and r.status == "queued" and r is not self._enrich_active
-            ]
-            internal_active = (
-                self._internal_active.to_summary()
-                if self._internal_active
-                and self._internal_active.status in _IN_FLIGHT_STATUSES
-                else None
-            )
-            internal_queue = [
-                r.to_summary()
-                for r in self._pending
-                if r._internal and r.status == "queued" and r is not self._internal_active
-            ]
-            history = list(self._history)
-        return {
-            "active": active,
-            "queue": queued,
-            "enrich_active": enrich_active,
-            "enrich_queue": enrich_queue,
-            "internal_active": internal_active,
-            "internal_queue": internal_queue,
-            "history": history,
-        }
-
-    def _finalize_run(self, run: Run) -> None:
-        with run._lock:
-            if run._finalized:
-                return
-            run._finalized = True
-        if not run._finished.is_set():
-            if run.ended_at is None:
-                run.ended_at = time.time()
-            run.mark_finished()
-            run.broadcast(
-                "done",
-                {
-                    "status": run.status,
-                    "exit_code": run.exit_code,
-                    "started_at": run.started_at,
-                    "ended_at": run.ended_at,
-                },
-            )
-        with self._lock:
-            if self._active is run:
-                self._active = None
-            if self._enrich_active is run:
-                self._enrich_active = None
-            if self._internal_active is run:
-                self._internal_active = None
-            if run in self._pending:
-                self._pending.remove(run)
-        self._unregister_active_process(run.id)
-        self._append_history(run.to_summary(), profile_id=run.profile_id)
-        self._persist_queue()
-        self._prune_runs_by_id()
-
-    def _worker_loop(self, lane: str = "fetcher") -> None:
-        lane_queue = self._lane_queue(lane)
-        while True:
-            try:
-                run = lane_queue.get()
-                if run is None:
-                    return
-                if not run._finished.is_set():
-                    with self._lock:
-                        self._set_lane_active(lane, run)
-                        if run.status == "queued":
-                            run.status = "launching"
-                    self._persist_queue()
-                    try:
-                        if run.status != "cancelled":
-                            self._execute(run)
-                    except Exception as exc:  # noqa: BLE001
-                        if not run.cancelled:
-                            run.status = "failed"
-                            run.exit_code = -1
-                            run.add_line("stderr", f"[server] worker error: {exc!r}")
-                self._finalize_run(run)
-            except Exception as exc:  # noqa: BLE001
-                print(f"[runs] worker loop error: {exc!r}", file=sys.stderr, flush=True)
-                time.sleep(0.5)
-
-    def _execute(self, run: Run) -> None:
-        if run.cancelled:
-            run.status = "cancelled"
-            run.exit_code = -1
-            return
-
-        argv = run.argv()
-        run.status = "launching"
-        run.started_at = time.time()
-        run.broadcast("status", {"status": run.status, "started_at": run.started_at})
-        run.add_line("stdout", f"$ {' '.join(argv)}")
-
-        try:
-            from auth.manager import subprocess_env_for_profile
-
-            env = subprocess_env_for_profile(run.profile_id)
-        except Exception as exc:
-            run.status = "failed"
-            run.exit_code = -1
-            run.add_line("stderr", f"[server] failed to build subprocess env: {exc!r}")
-            return
-
-        from baklog_fetcher_dispatch import apply_fetcher_env_mirror
-
-        apply_fetcher_env_mirror(argv, env)
-
-        # Run Popen on a launcher thread so a wedged CreateProcess can't wedge
-        # the queue worker. If the launch doesn't return within
-        # LAUNCH_TIMEOUT_SEC the run is marked failed and the worker moves on
-        # to the next queued item; late launches are terminated immediately.
-        launch_q: queue.Queue[tuple[str, Any]] = queue.Queue(maxsize=1)
-        launch_abandoned = threading.Event()
-
-        def _launch() -> None:
-            try:
-                p = popen_fetcher(  # noqa: S603 - argv is fixed in FETCHERS, not user input
-                    argv,
-                    # Run from repo root so the relative script path in argv resolves;
-                    # profile scoping is via BAKLOG_PROFILE in env + resolve_catalog_path,
-                    # not cwd (profiles/<id>/ holds data/cache, not the fetch_*.py scripts).
-                    cwd=str(data_root()),
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.STDOUT,
-                    bufsize=1,
-                    text=True,
-                    encoding="utf-8",
-                    errors="replace",
-                    env=env,
-                )
-            except BaseException as e:  # noqa: BLE001 - surface any launch failure to the worker
-                if not launch_abandoned.is_set():
-                    launch_q.put(("err", e))
-                return
-            if launch_abandoned.is_set():
-                if p.poll() is None and p.pid:
-                    _terminate_pid(p.pid)
-                run.add_line(
-                    "stderr",
-                    "[server] late launch after timeout — terminated stray subprocess",
-                )
-                return
-            launch_q.put(("ok", p))
-
-        threading.Thread(target=_launch, name=f"run-launch-{run.id}", daemon=True).start()
-        try:
-            tag, payload = launch_q.get(timeout=LAUNCH_TIMEOUT_SEC)
-        except queue.Empty:
-            launch_abandoned.set()
-            run.status = "failed"
-            run.exit_code = -1
-            run.add_line(
-                "stderr",
-                f"[server] subprocess launch did not return within {LAUNCH_TIMEOUT_SEC}s "
-                f"(likely Windows AppX Python activation deadlock); abandoning the launcher thread. "
-                f"Restart the server if subsequent runs also fail to start.",
-            )
-            return
-        if run.cancelled or run._finished.is_set():
-            if tag == "ok" and payload is not None and payload.poll() is None and payload.pid:
-                _terminate_pid(payload.pid)
-            run.status = "cancelled"
-            run.exit_code = -1
-            run.add_line("stderr", "[server] cancelled during launch")
-            return
-        if tag == "err":
-            exc = payload
-            if isinstance(exc, FileNotFoundError):
-                run.status = "failed"
-                run.exit_code = -1
-                run.add_line("stderr", f"[server] cannot launch: {exc}")
-                return
-            run.status = "failed"
-            run.exit_code = -1
-            run.add_line("stderr", f"[server] launch error: {exc!r}")
-            return
-        proc = payload
-
-        if run.cancelled:
-            if proc.poll() is None and proc.pid:
-                _terminate_pid(proc.pid)
-            run.status = "cancelled"
-            run.exit_code = -1
-            return
-
-        run._proc = proc
-        run.status = "running"
-        run.broadcast("status", {"status": run.status, "started_at": run.started_at})
-        if proc.pid:
-            self._register_active_process(run, proc.pid)
-        assert proc.stdout is not None
-
-        line_queue: queue.Queue[str | None] = queue.Queue()
-
-        def _reader() -> None:
-            try:
-                for raw in proc.stdout:
-                    line_queue.put(raw.rstrip("\n"))
-            finally:
-                line_queue.put(None)
-
-        threading.Thread(target=_reader, daemon=True).start()
-
-        last_line_at = time.monotonic()
-        last_stall_notice_at = 0.0
-        run_started_mono = time.monotonic()
-        max_run_sec = _max_run_seconds_for_key(run.key)
-        max_runtime_killed = False
-        reader_done = False
-
-        while not reader_done:
-            if run.cancelled:
-                break
-            try:
-                line = line_queue.get(timeout=STALL_POLL_SEC)
-            except queue.Empty:
-                line = "__POLL__"
-            now = time.monotonic()
-            if line == "__POLL__":
-                if proc.poll() is not None and line_queue.empty():
-                    break
-                elapsed = now - run_started_mono
-                if elapsed >= max_run_sec:
-                    run.add_line(
-                        "stderr",
-                        f"[server] exceeded maximum runtime ({int(max_run_sec)}s) — "
-                        f"force-killing PID {proc.pid}",
-                    )
-                    if proc.pid:
-                        _terminate_pid(proc.pid)
-                    try:
-                        proc.wait(timeout=TERMINATE_GRACE_SEC)
-                    except Exception:  # noqa: BLE001
-                        pass
-                    max_runtime_killed = True
-                    break
-                silent = now - last_line_at
-                if silent >= SILENT_STALL_KILL_SEC:
-                    run.add_line(
-                        "stderr",
-                        f"[server] no output for {int(silent)}s — force-killing PID {proc.pid}",
-                    )
-                    if proc.pid:
-                        _terminate_pid(proc.pid)
-                    try:
-                        proc.wait(timeout=TERMINATE_GRACE_SEC)
-                    except Exception:  # noqa: BLE001
-                        pass
-                    break
-                if silent >= STALL_FIRST_NOTICE_SEC and (
-                    last_stall_notice_at == 0.0
-                    or now - last_stall_notice_at >= STALL_REPEAT_SEC
-                ):
-                    sec = int(silent)
-                    run.add_line(
-                        "stderr",
-                        f"[server] no output for {sec}s — still running (PID {proc.pid})",
-                    )
-                    last_stall_notice_at = now
-                continue
-            if line is None:
-                reader_done = True
-                break
-            run.add_line("stdout", line)
-            last_line_at = now
-            last_stall_notice_at = 0.0
-
-        # On cancel the HTTP thread already issued _terminate_pid(); re-issue
-        # here in case that missed (Windows AppX Python can survive the first
-        # taskkill), then never block the worker indefinitely — a lingering
-        # zombie must not wedge the queue. We finalize after a bounded wait so
-        # the next queued run can start regardless.
-        if run.cancelled and proc.poll() is None:
-            if proc.pid:
-                _terminate_pid(proc.pid)
-            try:
-                proc.wait(timeout=TERMINATE_GRACE_SEC)
-            except Exception:  # noqa: BLE001 - subprocess.TimeoutExpired or platform variants
-                run.add_line(
-                    "stderr",
-                    f"[server] PID {proc.pid} did not exit after kill; "
-                    f"abandoning it and advancing the queue",
-                )
-        else:
-            try:
-                proc.wait(timeout=TERMINATE_GRACE_SEC)
-            except Exception:  # noqa: BLE001
-                if proc.pid:
-                    _terminate_pid(proc.pid)
-        run.exit_code = proc.returncode if proc.returncode is not None else -1
-        if max_runtime_killed:
-            run.status = "failed"
-            run.exit_code = -1
-        elif run.cancelled:
-            run.status = "cancelled"
-            run.add_line("stderr", "[server] cancelled")
-        elif proc.returncode == 0:
-            run.status = "done"
-        else:
-            run.status = "failed"
-        run._proc = None
-
-
-MANAGER = RunManager(restore_durable=False)
 
 # Pro-tier background scheduler (created/started in main(); None under pytest import).
 SCHEDULER: Any = None
